EOS-24415: Merge startup.py scripts from logrotate/config and logrotate/cron dir & move it to /startup.
